trade_ai/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0

# OpenAI & Agents
openai>=1.58.0
openai-agents>=0.0.7

# Vector Database
qdrant-client>=1.12.0

# Memory
mem0ai>=0.1.38

# Monitoring
langfuse>=2.56.0

# Web Search
tavily-python>=0.5.0

# AWS
boto3>=1.35.0

# PDF Processing
pymupdf>=1.24.0

# HTTP Client
httpx>=0.27.0

# Settings
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
